requirements.txt
recsim-ng
tensorflow==2.12.0
tensorflow-probability
edward2
tf-agents
gin-config
numpy
pandas
scipy
matplotlib
optuna